requirement.txt
alabaster==0.7.10
asn1crypto==0.24.0
astroid==1.6.3
attrs==18.1.0
Automat==0.7.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bleach==2.1.3
boto==2.48.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.2.2
cssselect==1.1.0
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
decorator==4.3.0
distributed==1.21.8
Django==2.0
django-cors-headers==3.1.0
django-redis==4.10.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
gevent==1.3.0
glob2==0.6
greenlet==0.4.13
heapdict==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.6
imagesize==1.0.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==6.0.1
numpy==1.14.3
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.4.5
py==1.5.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.9.0
pycryptodomex==3.7.2
pycurl==7.43.0.1
PyDispatcher==2.0.5
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==1.8.4
PyMySQL==0.9.3
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.5.1
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-alipay-sdk==1.10.1
python-dateutil==2.7.3
pytz==2018.4
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
qrcode==6.1
QtAwesome==0.4.4
qtconsole==4.5.4
QtPy==1.4.1
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
rope==0.10.7
ruamel-yaml==0.15.35
scipy==1.3.1
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
scrapy-redis==0.6.8
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.2.7
sqlparse==0.3.0
sympy==1.1.1
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==6.0.3
traitlets==4.3.2
Twisted==19.7.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zope.interface==4.6.0
